requirements.txt
tensorflow
numpy
python-chess
pygame
tqdm
pygame